requirements.txt
Django
requests
numpy
scikit-learn
pandas
python-dotenv
sqlparse